Ref: consolidate small factory modules
- Moved to tests/factories/__init__.py:
  - make_config
  - make_client
  - make_probe_client
  - make_persist_result
- Deleted config.py, dto.py, web.py from tests/factories/
- Existing submodule re-exports (models, payloads) unchanged
