+++ tests/web/test_errors.py
@@ -10,7 +10,7 @@ from bs4 import BeautifulSoup
 from fastapi import FastAPI, Request
 from fastapi.testclient import TestClient
 
-from tests.factories.web import make_client
+from tests.factories import make_client
 
 
 class TestError404Page:
