+++ tests/web/test_info_page.py
@@ -9,7 +9,7 @@ License: Apache-2.0
 from bs4 import BeautifulSoup
 
 from tests.factories.payloads import gen_image
-from tests.factories.web import make_client
+from tests.factories import make_client
 
 
 class TestInfoPageText:
