+++ tests/web/test_negotiate_routes.py
@@ -8,7 +8,7 @@ License: Apache-2.0
 
 import pytest
 
-from tests.factories.web import make_client
+from tests.factories import make_client
 
 
 class TestShortcutRoute:
