+++ tests/web/test_upload_htmx.py
@@ -8,7 +8,7 @@ License: Apache-2.0
 
 from bs4 import BeautifulSoup
 
-from tests.factories.web import make_client
+from tests.factories import make_client
 
 
 class TestHtmxUploadSuccess:
